#62 fix: OAuth window popup 닫힘 현상 해결 및 SignInLoadingPage 구현
