+++ src/components/auth/NaverSignInButton.tsx
@@ -14,20 +14,9 @@ export default function NaverSignInButton() {
   const onNaverSignIn = async () => {
     const res = await postOAuthUrl("naver");
 
-    // This is for development. Remove this.
-    const tempURL = new URL(res.data.authURL);
-    tempURL.searchParams.set(
-      "redirect_uri",
-      "http://localhost:5173/signin?provider=naver"
-    );
-    const url =
-      process.env.NODE_ENV === "development"
-        ? tempURL.toString()
-        : res.data.authURL;
-
     if (res.code === HTTPSTATUS.success) {
       const oAuthPopUpWindow = openPopUpWindow(
-        url, // This is for development. Change this to res.data.authURL.
+        res.data.authURL,
         "naverOAuth",
         500,
         600
@@ -39,10 +28,8 @@ export default function NaverSignInButton() {
     }
   };
 
-  // TODO: 로고 야매로 잘라서 모카가 만들어주신 피그마에서 따서 해야할듯
   return (
     <StyledNaverSignInButton type="button" onClick={onNaverSignIn}>
-      {/* <img src={naverLoginButtonImg} alt="네이버 로그인" /> */}
       <img src={naverLogo} alt="네이버 로고" />
       <p>네이버 로그인</p>
     </StyledNaverSignInButton>
