+++ src/pages/SignInPage.tsx
@@ -54,42 +54,33 @@ export default function SignInPage() {
     signInMutate({ email, password });
   };
 
-  // Handle Google, Kakao, Naver Redirect (receive OAuth provider, auth code and state) received in popup.
-  useEffect(() => {
-    const urlParams = new URLSearchParams(window.location.search);
-    const provider = urlParams.get("provider");
-    const authCode = urlParams.get("code");
-    const state = urlParams.get("state");
-
-    if (!provider || !authCode || !state) return; // TODO: handle error
-
-    // Received data in the popup window from the OAuth provider and send it to the original window.
-    if (provider === "google" || provider === "kakao" || provider === "naver") {
-      // Send OAuth provider, auth code and state to the original window.
-      window.opener.postMessage({ provider, authCode, state }, CLIENT_URL);
-    }
-  }, [oAuthSignInMutate]);
-
   // Receive OAuth provider, auth code, state in original window from popup window.
   // Only used by Kakao and Naver Login.
   useEffect(() => {
     if (!popUpWindow) return;
 
-    const closeWindowMessageHandler = async (e: MessageEvent) => {
+    let shouldClosePopUp = false;
+
+    const closePopUpMessageHandler = async (e: MessageEvent) => {
       if (e.origin === CLIENT_URL) {
         const { provider, authCode, state } = e.data;
 
         if (provider && authCode && state) {
           await oAuthSignInMutate({ provider, authCode, state });
         }
-        closePopUpWindow();
+
+        shouldClosePopUp = true;
       }
     };
 
-    window.addEventListener("message", closeWindowMessageHandler);
+    window.addEventListener("message", closePopUpMessageHandler);
 
     return () => {
-      window.removeEventListener("message", closeWindowMessageHandler);
+      window.removeEventListener("message", closePopUpMessageHandler);
+
+      if (shouldClosePopUp) {
+        closePopUpWindow();
+      }
     };
   }, [closePopUpWindow, oAuthSignInMutate, popUpWindow]);
